Move entitlement checkers to policy package.
Maintain parity with main branch to ease backporting.
